Make Stripe amount BigDecimal after it's divided by 100 (#49)

+++ src/main/scala/model/stripe/StripeChargeSuccess.scala
@@ -3,9 +3,9 @@ package model.stripe
 import com.stripe.model.Charge
 import io.circe.generic.JsonCodec
 
-@JsonCodec case class StripeChargeSuccess private (currency: String, amount: Long)
+@JsonCodec case class StripeChargeSuccess private (currency: String, amount: BigDecimal)
 
 object StripeChargeSuccess {
   def fromCharge(charge: Charge): StripeChargeSuccess =
-    StripeChargeSuccess(charge.getCurrency, charge.getAmount / 100)
+    StripeChargeSuccess(charge.getCurrency, BigDecimal(charge.getAmount) / 100)
 }
